+++ changelog.md
@@ -13,6 +13,7 @@ Codion Change Log
 - EntitySearchModel.stringFunction() now final and no longer Value based.
 - EntitySearchModel.stringFunction() renamed stringFactory().
 - EntitySearchModel.separator() now final and no longer Value based.
+- EntitySearchModel.description() removed.
 ### is.codion.swing.common.ui
 - FilterTableCellRenderer.UISettings.selectionForeground() added.
 - DefaultFilterTableCellRenderer uses UISettings.selectionForeground().
@@ -29,6 +30,7 @@ Codion Change Log
 - EntitySearchField.selectorFactory now final, selectorFactory() removed.
 - EntitySearchField.searchIndicator now final, searchIndicator() removed.
 - EntityComponents.supports() removed, unused.
+- EntitySearchField now displays the selected entities in a multiline tooltip instead of the model description.
 
 ## 0.18.24
 ## is.codion.common.core